refactor(coral): Extract renderForm from Form.test.tsx
